242: rename util file with group code to group.ts
