+++ src/App.tsx
@@ -7,8 +7,7 @@ import { createStore } from "solid-js/store";
 import "./styles/setup.css";
 import { Background } from "./components/Background/Background.tsx";
 import { TARGET, TargetProvider } from "./contexts/target.tsx";
-import { LINE_TYPE, type Node } from "./utils/zon";
-import { createTree } from "./utils/zon/tokei-adapter.ts";
+import { LINE_TYPE, type Node, createTree } from "./utils/zon";
 import HomePage from "./pages/HomePage/HomePage.tsx";
 import ReportPage from "./pages/ReportPage/ReportPage.tsx";
 import { Languages } from "./utils/tokei.ts";
