+++ src/utils/zon/index.ts
@@ -2,5 +2,6 @@ export * from "./types.ts";
 export * from "./node.ts";
 export * from "./path.ts";
 export * from "./lineType.ts";
-export * from "./tree.ts";
+export * from "./group.ts";
 export * from "./color.ts";
+export * from "./tokei-adapter.ts";
